backend/requirements.txt
fastapi
sqlalchemy
uvicorn
pymysql
python-multipart
pydantic
pyjwt
passlib[bcrypt]
email_validator